backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.12
pandas>=2.2.3
openpyxl==3.1.5
google-genai>=1.0.0
python-dotenv==1.0.1
pdfplumber>=0.11.0
ppi-client>=1.2.4
requests>=2.33.0
python-telegram-bot>=21.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
